requirements.txt
appdirs==1.4.0
decorator==4.0.11
ipython==5.1.0
ipython-genutils==0.1.0
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.1.10
PySide==1.2.4
PyYAML==3.12
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
wcwidth==0.1.7
